client and server options abstracted
